OpenBSD 6.7 errata 001 (001_wscons.patch.sig)
CVSROOT:        /cvs
Module name:    src
Changes by:     kettenis@cvs.openbsd.org        2020/05/10 14:50:55

Modified files:
        sys/dev/wscons : wsdisplay.c wsemulconf.c

Log message:
Fix two out-of-bounds array accesses in ioctl code pats.  One found by
johnc@armadilloaerospace.com and another one spotted by matthieu@.

ok benno@, matthieu@, deraadt@

+++ gnu/usr.bin/binutils-2.17/bfd/ecoff.c
@@ -1878,7 +1878,7 @@ _bfd_ecoff_sizeof_headers (bfd *abfd, bfd_boolean reloc ATTRIBUTE_UNUSED)
   ret = (bfd_coff_filhsz (abfd)
 	 + bfd_coff_aoutsz (abfd)
 	 + c * bfd_coff_scnhsz (abfd));
-  return BFD_ALIGN (ret, 16);
+  return (int) BFD_ALIGN (ret, 16);
 }
 
 /* Get the contents of a section.  */
@@ -3050,7 +3050,7 @@ _bfd_ecoff_write_armap (bfd *abfd,
      linker just checks the archive name; the GNU linker may check the
      date.  */
   stat (abfd->filename, &statbuf);
-  sprintf (hdr.ar_date, "%ld", (long) (statbuf.st_mtime + 60));
+  sprintf (hdr.ar_date, "%lld", (long long) (statbuf.st_mtime + 60));
 
   /* The DECstation uses zeroes for the uid, gid and mode of the
      armap.  */
