+++ include/protocols/rwhod.h
@@ -1,3 +1,4 @@
+/*	$OpenBSD: rwhod.h,v 1.3 1997/09/21 10:46:06 niklas Exp $	*/
 /*	$NetBSD: rwhod.h,v 1.3 1994/10/26 00:56:51 cgd Exp $	*/
 
 /*
@@ -12,11 +13,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the University of
- *	California, Berkeley and its contributors.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -44,21 +41,21 @@
 struct	outmp {
 	char	out_line[8];		/* tty name */
 	char	out_name[8];		/* user id */
-	long	out_time;		/* time on */
+	int32_t	out_time;		/* time on */
 };
 
 struct	whod {
 	char	wd_vers;		/* protocol version # */
 	char	wd_type;		/* packet type, see below */
 	char	wd_pad[2];
-	int	wd_sendtime;		/* time stamp by sender */
-	int	wd_recvtime;		/* time stamp applied by receiver */
+	int32_t	wd_sendtime;		/* time stamp by sender */
+	int32_t	wd_recvtime;		/* time stamp applied by receiver */
 	char	wd_hostname[32];	/* hosts's name */
-	int	wd_loadav[3];		/* load average as in uptime */
-	int	wd_boottime;		/* time system booted */
+	int32_t	wd_loadav[3];		/* load average as in uptime */
+	int32_t	wd_boottime;		/* time system booted */
 	struct	whoent {
-		struct	outmp we_utmp;	/* active tty info */
-		int	we_idle;	/* tty idle time */
+		struct		outmp we_utmp;	/* active tty info */
+		int32_t		we_idle;	/* tty idle time */
 	} wd_we[1024 / sizeof (struct whoent)];
 };
